feat: gcd_large_numbers

+++ implementation-euclidean-gcd/src/main/kotlin/com/currenjin/Euclidean.kt
@@ -5,12 +5,12 @@ import kotlin.math.abs
 class Euclidean {
     companion object {
         fun gcd(
-            n1: Int,
-            n2: Int,
-        ): Int {
+            n1: Long,
+            n2: Long,
+        ): Long {
             val x = abs(n1)
             val y = abs(n2)
-            if (y == 0) return x
+            if (y == 0L) return x
             return gcd(y, x % y)
         }
     }
